qna detail 페이지의 댓글목록, 페이징(Session만 들어오면 완료), validation(reserve, qna) 작업

+++ src/main/java/com/lec/spring/domain/ReserveWrite.java
@@ -14,9 +14,9 @@ import java.time.LocalDateTime;
 public class ReserveWrite { //여기에서는 name, phone, regdate, reserve_date 를 받아야함.
     private Long id;
     private String name;
-    private int phone;
+    private Long phone;
     private LocalDateTime regdate;
-    private int date;
+    private Long date;
     private int leisure_id;
     private int user_id;
